refactor: remove database worker re-export
