--- src/lorairo/gui/workers/database_worker.py
@@ -1,20 +1,0 @@
-"""データベース関連ワーカー - 後方互換re-exportモジュール。
-
-各ワーカーは個別ファイルに分離済み:
-- registration_worker.py: DatabaseRegistrationWorker, DatabaseRegistrationResult
-- search_worker.py: SearchWorker, SearchResult
-- thumbnail_worker.py: ThumbnailWorker, ThumbnailLoadResult
-"""
-
-from .registration_worker import DatabaseRegistrationResult, DatabaseRegistrationWorker
-from .search_worker import SearchResult, SearchWorker
-from .thumbnail_worker import ThumbnailLoadResult, ThumbnailWorker
-
-__all__ = [
-    "DatabaseRegistrationResult",
-    "DatabaseRegistrationWorker",
-    "SearchResult",
-    "SearchWorker",
-    "ThumbnailLoadResult",
-    "ThumbnailWorker",
-]
